requirements.txt
openai>=1.30.0
anthropic>=0.30.0
datasets>=2.18.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
tqdm>=4.66.0
nltk>=3.8.1
rake-nltk>=1.0.6
unbabel-comet>=2.2.2
pyyaml>=6.0.1
scipy>=1.11.0